requirements.txt
apache-beam==2.38.0
arrow==1.2.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
cloudpickle==2.0.0
colorama==0.4.4
crcmod==1.7
debugpy==1.6.0
decorator==5.1.1
dill==0.3.1.1
docopt==0.6.2
entrypoints==0.4
executing==0.8.3
fastavro==1.4.11
fasteners==0.17.3
fastjsonschema==2.15.3
google-api-core==1.31.5
google-apitools==0.5.31
google-auth==1.35.0
google-cloud-bigquery==2.34.3
google-cloud-bigquery-storage==2.13.1
google-cloud-bigtable==1.7.1
google-cloud-core==1.7.2
google-cloud-datastore==1.15.4
google-cloud-dlp==3.6.2
google-cloud-language==1.3.1
google-cloud-pubsub==2.12.0
google-cloud-pubsublite==1.4.2
google-cloud-recommendations-ai==0.2.0
google-cloud-spanner==1.19.2
google-cloud-videointelligence==1.16.2
google-cloud-vision==1.0.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.4
grpcio==1.44.0
grpcio-gcp==0.2.2
grpcio-status==1.44.0
hdfs==2.7.0
httplib2==0.19.1
idna==3.3
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
jsonschema==4.4.0
jupyter-client==7.3.0
jupyter-core==4.10.0
matplotlib-inline==0.1.3
nbformat==5.3.0
nest-asyncio==1.5.5
numpy==1.22.3
oauth2client==4.1.3
orjson==3.6.8
overrides==6.1.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
pigar==0.10.0
prompt-toolkit==3.0.29
proto-plus==1.20.3
protobuf==3.20.1
psutil==5.9.0
pure-eval==0.2.2
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
Pygments==2.12.0
pymongo==3.12.3
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pyzmq==22.3.0
requests==2.27.1
rsa==4.8
six==1.16.0
stack-data==0.2.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing-utils==0.1.0
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
